# OEUPB-Docs/specs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OE UPB - Observatorio de Egresados",
    "description": "API RESTful para el procesamiento de egresados UPB.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/carga/excel": {
      "post": {
        "tags": [
          "Carga de Datos"
        ],
        "summary": "Procesar Excel",
        "operationId": "procesar_excel_api_carga_excel_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_procesar_excel_api_carga_excel_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaResponse"
                }
              }
            }
          },
          "400": {
            "description": "El archivo no es .xlsx o su contenido no puede leerse"
          },
          "403": {
            "description": "El usuario no es coordinador o no tiene sede asignada"
          },
          "409": {
            "description": "El archivo es idéntico a la versión vigente del mismo alcance"
          },
          "422": {
            "description": "El momento o los parámetros no son válidos"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/carga/historial": {
      "get": {
        "tags": [
          "Carga de Datos"
        ],
        "summary": "Get Historial",
        "operationId": "get_historial_api_carga_historial_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistorialCargaItem"
                  },
                  "type": "array",
                  "title": "Response Get Historial Api Carga Historial Get"
                }
              }
            }
          },
          "403": {
            "description": "El usuario no es coordinador o no tiene sede asignada"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/carga/archivo/{carga_id}": {
      "delete": {
        "tags": [
          "Carga de Datos"
        ],
        "summary": "Eliminar Carga",
        "operationId": "eliminar_carga_api_carga_archivo__carga_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EliminarCargaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeResponse"
                }
              }
            }
          },
          "403": {
            "description": "El usuario no es coordinador o no tiene sede asignada"
          },
          "404": {
            "description": "La carga no existe dentro de la sede autorizada"
          },
          "409": {
            "description": "La carga ya no está vigente"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_procesar_excel_api_carga_excel_post": {
        "properties": {
          "momento": {
            "type": "integer",
            "title": "Momento",
            "description": "Momento 0, 1 o 5"
          },
          "anio": {
            "type": "integer",
            "title": "Anio",
            "description": "Año de grado o cohorte (1900-2200)"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Archivo Excel en formato .xlsx"
          }
        },
        "type": "object",
        "required": [
          "momento",
          "anio",
          "file"
        ],
        "title": "Body_procesar_excel_api_carga_excel_post"
      },
...
      "CargaResponse": {
        "properties": {
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "errores": {
            "items": {
              "$ref": "#/components/schemas/ErrorFila"
            },
            "type": "array",
            "title": "Errores",
            "default": []
          },
          "carga_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carga Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado"
          },
          "registros": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registros"
          }
        },
        "type": "object",
        "required": [
          "mensaje"
        ],
        "title": "CargaResponse"
      },
...
      "EliminarCargaRequest": {
        "properties": {
          "motivo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 10,
            "title": "Motivo"
          }
        },
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "EliminarCargaRequest"
      },
...
      "ErrorFila": {
        "properties": {
          "fila": {
            "type": "integer",
            "title": "Fila"
          },
          "columna": {
            "type": "string",
            "title": "Columna"
          },
          "error": {
            "type": "string",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "fila",
          "columna",
          "error"
        ],
        "title": "ErrorFila"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistorialCargaItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "momento": {
            "type": "integer",
            "title": "Momento"
          },
          "anio": {
            "type": "integer",
            "title": "Anio"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "nombre_archivo": {
            "type": "string",
            "title": "Nombre Archivo"
          },
          "fecha_carga": {
            "type": "string",
            "title": "Fecha Carga"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "registros": {
            "type": "integer",
            "title": "Registros"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "momento",
          "anio",
          "nombre",
          "nombre_archivo",
          "fecha_carga",
          "version",
          "registros",
          "estado"
        ],
        "title": "HistorialCargaItem"
      },
...
      "MensajeResponse": {
        "properties": {
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          }
        },
        "type": "object",
        "required": [
          "mensaje"
        ],
        "title": "MensajeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}